Add variables

+++ src/main.rs
@@ -19,6 +19,13 @@ struct Args {
 
     #[arg(short, long)]
     output_path: String,
+
+    #[arg(
+        long,
+        default_value = "false",
+        help = "Disables the assembler. Writes the assembly file to disk."
+    )]
+    no_assemble: bool,
 }
 
 fn main() {
@@ -29,36 +36,42 @@ fn main() {
     let lexer = Lexer::default();
     let tokens = lexer.tokenize(text);
 
-    let parser = Parser::default();
-    let exit_node = parser.parse(tokens);
+    let mut parser = Parser::default();
+    let root_node = parser.parse(tokens);
 
-    let asm = compiler::compile(exit_node);
+    let asm = compiler::compile(root_node);
     write(format!("{}.asm", args.output_path), asm).unwrap();
 
-    Command::new("nasm")
-        .args([
-            "-felf64",
-            format!("{}.asm", args.output_path).as_str(),
-            "-o",
-            format!("{}.o", args.output_path).as_str(),
-        ])
-        .status()
-        .unwrap();
+    if !args.no_assemble {
+        Command::new("nasm")
+            .args([
+                "-felf64",
+                format!("{}.asm", args.output_path).as_str(),
+                "-o",
+                format!("{}.o", args.output_path).as_str(),
+            ])
+            .status()
+            .unwrap();
 
-    Command::new("ld")
-        .args([
-            format!("{}.o", args.output_path).as_str(),
-            "-o",
-            args.output_path.as_str(),
-        ])
-        .status()
-        .unwrap();
+        Command::new("ld")
+            .args([
+                format!("{}.o", args.output_path).as_str(),
+                "-o",
+                args.output_path.as_str(),
+            ])
+            .status()
+            .unwrap();
 
-    remove_file(format!("{}.asm", args.output_path)).unwrap();
-    remove_file(format!("{}.o", args.output_path)).unwrap();
+        remove_file(format!("{}.o", args.output_path)).unwrap();
+        remove_file(format!("{}.asm", args.output_path)).unwrap();
+    }
 
     println!(
         "Compilation successful. File outputted at: {}",
-        args.output_path
+        if args.no_assemble {
+            format!("{}.asm", args.output_path)
+        } else {
+            args.output_path
+        },
     );
 }
